update dependency image loader for bug fix, and fix a bug in legacy qupath export

+++ src/main/java/ch/epfl/biop/atlas/aligner/SliceSources.java
@@ -1031,7 +1031,9 @@ public class SliceSources {
                     if (mp.getAtlas() instanceof AllenBrainAdultMouseAtlasCCF2017Command) {
                         // Save it the old fashioned way
                         f = new File(dataEntryFolder, "ABBA-RoiSet.zip");
-                        Files.delete(Paths.get(f.getAbsolutePath()));
+                        if (f.exists()) {
+                            Files.delete(Paths.get(f.getAbsolutePath()));
+                        }
                         Files.copy(Paths.get(ijroisfile.f.getAbsolutePath()),Paths.get(f.getAbsolutePath()));
                     }
                     //----------------- LEGACY
